+++ MACS2/cStatistics.h
@@ -1,9 +1,9 @@
-double log10_poisson_cdf ( unsigned int n, double lam, short lower );
-double log10_poisson_cdf_P_large_lambda ( unsigned int k, double lbd );
-double log10_poisson_cdf_Q_large_lambda ( unsigned int k, double lbd );
-double chi2_k1_cdf ( double x );
-double chi2_k2_cdf ( double x );
-double chi2_k4_cdf ( double x );
-double log10_chi2_k1_cdf ( double x );
-double log10_chi2_k2_cdf ( double x );
-double log10_chi2_k4_cdf ( double x );
+float log10_poisson_cdf ( unsigned int n, float lam, short lower );
+float log10_poisson_cdf_P_large_lambda ( unsigned int k, float lbd );
+float log10_poisson_cdf_Q_large_lambda ( unsigned int k, float lbd );
+float chi2_k1_cdf ( float x );
+float chi2_k2_cdf ( float x );
+float chi2_k4_cdf ( float x );
+float log10_chi2_k1_cdf ( float x );
+float log10_chi2_k2_cdf ( float x );
+float log10_chi2_k4_cdf ( float x );
